docs: 실행화면 이미지 추가
