+++ ios-wanted-VideoRecorder/VideoRecorderTests/Tests/VideoUseCaseTests.swift
@@ -1,0 +1,27 @@
+//
+//  VideoUseCaseTests.swift
+//  VideoRecorderTests
+//
+//  Created by 강민수 on 2023/06/10.
+//
+
+@testable import ios_wanted_VideoRecorder
+import AVKit
+import XCTest
+
+final class VideoUseCaseTests: XCTestCase {
+
+    override func setUpWithError() throws {
+    }
+
+    override func tearDownWithError() throws {
+    }
+
+    func testExample() throws {
+        // This is an example of a functional test case.
+        // Use XCTAssert and related functions to verify your tests produce the correct results.
+        // Any test you write for XCTest can be annotated as throws and async.
+        // Mark your test throws to produce an unexpected failure when your test encounters an uncaught error.
+        // Mark your test async to allow awaiting for asynchronous code to complete. Check the results with assertions afterwards.
+    }
+}
